10-8

+++ Pixiv.cs
@@ -22,15 +22,15 @@ namespace PixivDaily
         }
 
 
-        public void get_pixiv_images(string date)
+        public void Get_pixiv_images(string date)
         {
             try
             {
                 DateTime start = DateTime.Now;
-                int pagecount = 1;//控制排行拉取榜页数，最高10
+                int pagecount = 10;//控制排行拉取榜页数，最高10
                 int page = 1;
 
-                string savepath = $"{Application.StartupPath}\\original_images\\{date}";
+                string savepath = $@"{Application.StartupPath}\original_images\{date}";
                 if (!System.IO.Directory.Exists(savepath))
                 {
                     System.IO.Directory.CreateDirectory(savepath);
@@ -40,24 +40,22 @@ namespace PixivDaily
 
                 }
                 List<string> imgurls = new List<string>();
-
-                //Regex reg = new Regex(@"(?<=data-filter=""thumbnail-filter lazy-image""data-src="")(.+?\.jpg)(?=""data-type=""illust"")");
-                //Regex reg = new Regex(@"(?<=""new""(.+?)data-filter=""thumbnail-filter lazy-image""data-src="")(.+?\.jpg)(?=""data-type=""illust"")");
-                Regex reg = new Regex(@"(?<=class=""new""((?!之前 #\d).)+data-filter=""thumbnail-filter lazy-image""data-src="")(.+?\.jpg)(?=""data-type=""illust"")");
+                Regex reg = new Regex(@"(?<=class=""new""((?!前日).)+data-filter=""thumbnail-filter lazy-image""data-src="")(.+?\.jpg)(?=""data-type=""illust"")");
+                //Regex reg = new Regex(@"(?<=class=""new""((?!之前 #\d).)+data-filter=""thumbnail-filter lazy-image""data-src="")(.+?\.jpg)(?=""data-type=""illust"")");
                 while (page <= pagecount )
                 {
-                    //string html = get_html($"https://www.pixiv.net/ranking.php?mode=daily&content=illust&p={page}&date={date}");
-                    string html = File.ReadAllText("1.html");//读文件模拟返回页面
+                    string html = Get_html($"https://www.pixiv.net/ranking.php?mode=daily&content=illust&p={page}&date={date}");
+                    //string html = File.ReadAllText("1.html");//读文件模拟返回页面
 
-                    List<string> imgurl = new List<string>();
+                    //List<string> imgurl = new List<string>();
                     foreach (var url in reg.Matches(html))
                     {
-                        imgurl.Add(url.ToString());
+                        imgurls.Add(url.ToString());
                     }
-                    imgurls.AddRange(imgurl);
+                    //imgurls.AddRange(imgurl);
                     page += 1;
                 }
-                save_image(imgurls, savepath);
+                Save_image(imgurls, savepath);
             }
             catch (Exception e)
             {
@@ -66,33 +64,37 @@ namespace PixivDaily
             }
         }
 
-        public void save_image(List<string> urls, string save_path)
+        public void Save_image(List<string> urls, string save_path)
         {
+            int i = 0;
             foreach (string _url in urls)
             {
                 string url = _url.Replace("c/240x480/img-master", "img-original");
                 url = url.Replace("_master1200", "");
                 url = url.Replace(".jpg", ".png");
-                string image_path = $"{save_path}\\{get_pixiv_id(url)}";
+                string image_path = $"{save_path}\\{Get_pixiv_id(url)}";
                 try
                 {
                     DownloadFile(url, image_path + ".png");
-                    SetText(get_pixiv_id(url)+"\r\n");
+                    SetText(Get_pixiv_id(url)+"\r\n");
                 }
                 catch (Exception e)
                 {
                     if (e.Message== "远程服务器返回错误: (404) 未找到。")
                     {
-                        SetText($"{e.Message }{ get_pixiv_id(url)}，使用.JPG尝试\r\n");
+                        SetText($"{e.Message }{ Get_pixiv_id(url)}，使用.JPG尝试\r\n");
                         url = url.Replace(".png", ".jpg");
                         DownloadFile(url, image_path + ".jpg");
-                        SetText(get_pixiv_id(url) + "\r\n");
+                        
+                        SetText(Get_pixiv_id(url) + "\r\n");
                     }
                     else
                     {
                         throw e;
                     }
                 }
+                i++;
+                BGWorker.ReportProgress(i,urls.Count);
             }
         }
 
@@ -101,6 +103,7 @@ namespace PixivDaily
             try
             {
                 if (File.Exists(FilePath)) return;
+                if (File.Exists(FilePath.Replace(".png", ".jpg"))) return;
                 string tempFile = Application.StartupPath + @"\temp\downloading.temp";
                 File.Delete(tempFile);
                 HttpWebRequest request = (HttpWebRequest)WebRequest.Create(url);
@@ -108,19 +111,33 @@ namespace PixivDaily
                 request.UserAgent = "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/69.0.3497.92 Safari/537.36";
                 request.AllowAutoRedirect = false;
                 request.Timeout = 300000;
-                request.Referer = get_referer(url);
+                request.Referer = Get_referer(url);
                 HttpWebResponse response = (HttpWebResponse)request.GetResponse();
                 Stream sr = response.GetResponseStream();
-                
-                byte[] bArr = new byte[1024];
-                int size = sr.Read(bArr, 0, (int)bArr.Length);
-                FileStream fs = new FileStream(tempFile, FileMode.Append, FileAccess.Write, FileShare.ReadWrite);
-                while (size > 0)
+
+                using (Stream fs = new FileStream(tempFile, FileMode.Append, FileAccess.Write, FileShare.ReadWrite))
                 {
-                    fs.Write(bArr, 0, size);
-                    size = sr.Read(bArr, 0, (int)bArr.Length);
+                    long num = 0L;
+                    byte[] array = new byte[1024];
+                    int size = sr.Read(array, 0, array.Length);
+                    while (size > 0)
+                    {
+                        num = (long)size + num;
+                        Application.DoEvents();
+                        fs.Write(array, 0, size);
+                        size = sr.Read(array, 0, array.Length);
+                        Application.DoEvents();
+                    }
                 }
-                fs.Close();
+                //byte[] bArr = new byte[2048];
+                //int size = sr.Read(bArr, 0, (int)bArr.Length);
+                //FileStream fs = new FileStream(tempFile, FileMode.Append, FileAccess.Write, FileShare.ReadWrite);
+                //while (size > 0)
+                //{
+                //    fs.Write(bArr, 0, size);
+                //    size = sr.Read(bArr, 0, (int)bArr.Length);
+                //}
+                //fs.Close();
                 sr.Close();
                 File.Move(tempFile, FilePath);
             }
@@ -130,18 +147,18 @@ namespace PixivDaily
             }
         }
 
-        public string get_pixiv_id(string url)
+        public string Get_pixiv_id(string url)
         {
             Regex reg = new Regex(@"(\d+)(?=_p0)");
             return reg.Match(url).Value;
         }
 
-        public string get_referer(string url)
+        public string Get_referer(string url)
         {
             string reference = "https://www.pixiv.net/member_illust.php?mode=medium&illust_id=";
-            return reference + get_pixiv_id(url)+ "&page=0";
+            return reference + Get_pixiv_id(url)+ "&page=0";
         }
-        public string get_html(string url)
+        public string Get_html(string url)
         {
             HttpWebRequest request = (HttpWebRequest)WebRequest.Create(url);
             request.Method = "GET";
@@ -174,32 +191,40 @@ namespace PixivDaily
 
         private void BGWorker_DoWork(object sender, DoWorkEventArgs e)
         {
-            get_pixiv_images(dateTimePicker1.Value.ToString("yyyyMMdd"));
+            Get_pixiv_images(dateTimePicker1.Value.ToString("yyyyMMdd"));
             //System.Threading.Thread.Sleep(10);
             
         }
 
         private void BGWorker_ProgressChanged(object sender, ProgressChangedEventArgs e)
         {
-            TextBox.AppendText(e.UserState.ToString()+"%\r\n");
+            int total = int.Parse(e.UserState.ToString());
+            int curr = e.ProgressPercentage;
+            progressBar1.Maximum = total;
+            progressBar1.Value = e.ProgressPercentage;
+            label1.Text = (curr *100/total).ToString();
         }
 
         private void BGWorker_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
         {
             TextBox.AppendText("Completed");
-            button_images.Enabled = true;
+            Button_images.Enabled = true;
         }
 
-        private void button_BGW_Click(object sender, EventArgs e)
+        private void Button_BGW_Click(object sender, EventArgs e)
         {
             if (BGWorker.IsBusy) return;
-            this.button_images.Enabled = false;
+            this.Button_images.Enabled = false;
             BGWorker.RunWorkerAsync();
         }
 
-        private void button_images_Click(object sender, EventArgs e)
+        private void Button_images_Click(object sender, EventArgs e)
         {
-            get_pixiv_images(dateTimePicker1.Value.ToString("yyyyMMdd"));
+            //Get_pixiv_images(dateTimePicker1.Value.ToString("yyyyMMdd"));
+            string urll = "https://i.pximg.net/c/240x480/img-master/img/2019/08/07/00/00/07/76116204_p0_master1200.jpg";
+            string url = urll.Replace("c/240x480/img-master", "img-original");
+            url = url.Replace("_master1200", "");
+            DownloadFile(url,Application.StartupPath+"\\1.jpg");
         }
     }
 }
